Fix tabindex logic; disable arrow key navigation
This plugin does a lot to override built-in browser behavior, some of
which is a good idea and somme of which definitely is not.  Overriding
the behavior of Tab and Shift+Tab is not a good idea.

This is related to the code used to navigate between FAQs using the
arrow keys.  Let's just remove that, and restore the default browser
behavior there too.